% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr_functions.R
\name{getHTTrChemPlates}
\alias{getHTTrChemPlates}
\title{Get the  HTTr plate_ids for a chemical}
\usage{
getHTTrChemPlates(epa_sample_id)
}
\arguments{
\item{epa_sample_id}{A valid epa_sample_id (e.g. from searchHTTrChem)}
}
\value{
A named list 
   plates: all plate_ids with chemical
}
\description{
Get the  HTTr plate_ids for a chemical
}
\examples{
Get plates for TP0001718A17
 getHTTrChemPlates('TP0001718A17')      
}
