% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr_functions.R
\name{getHTTrCountData}
\alias{getHTTrCountData}
\title{Get the raw HTTr count data by epa_sample_id}
\usage{
getHTTrCountData(epa_sample_id, probe_filter = "av0_10")
}
\arguments{
\item{epa_sample_id}{A valid epa_sample_id from searchHTTrChem}

\item{probe_filter}{removes probes that don't meet minimum requirements
mx0_5:  probe maximum greater than 5
av0_10: probe average greater than 10
md0_20: probe median greater than 20}
}
\value{
a named list with two dataframes
   treatments: all treatment factors (columns) x samples (rows)
   counts: all probe counts (columns) x samples (rows)
}
\description{
Get the raw HTTr count data by epa_sample_id
}
\examples{
Get count data for TP0001718N03 and keep probes with average > 100
 getHTTrCountData('TP0001718N03',probe_filter='av0_100')         
}
