% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr_functions.R
\name{getHTTrDEG}
\alias{getHTTrDEG}
\title{Gets the effect of each chemical treatment on probes}
\usage{
getHTTrDEG(sid, mncnt0 = NULL, p0 = NULL, l2fc0 = NULL,
  pivot_on = NULL, pivot_func = NULL, svc = SVC)
}
\arguments{
\item{sid}{A valid dsstox_sid (e.g. from searchHTTrChem)}

\item{p0}{the the unadjusted p-value cut-off for each probe (optional)}

\item{l2fc0}{the min abs(L2FC) threshold (optional)}

\item{pivot_on}{whether to pivot and how. Values
NULL: no pivot provide skinny results
gene: pivot on gene symbol 
probe_id: pivot on probe_id}

\item{mincnt0}{the min average raw count of the probe across reps (optional)}

\item{pivot_func:}{if pivot_on=='gene' then how to aggregate l2fc
max: take the max +ve or min -ve
mean: take the average}
}
\value{
a named list with two dataframes
   l2fc:  all L2FC data by probe_id/genen (columns) x treatment groups (rows)
   index: the index columns in the L2FC dataframe
}
\description{
This function returns the results from DESeq2 analysis of the raw                 
 count data by using the replicates for each sample across multiple plates
 as well as the DMSO conrols for each treatment group (conc).
 Just calling this with the sid wil return a long skinny format df. The skinny
 df has all the information for filtering the results. 
 Optionally, these results can be filtered by p-values, minimum average count
 l2fc magnitude and also pivoted to create wide df (more suitable for analysis)
}
\details{
Note:
 a) If the chemical was run mulitple times then there will be multiple sample ids!
 b) If filters are used to exclude probes by any criterion then the missing value is replaced with 0.0 (naught) 
 c) Keep the skinny format without pivoting if you want to see all the data
}
\examples{
  Get all the probes with differential expression information for DTXSID7020182
    X <- getHTTrDEG('DTXSID7020182')

  ... above chemical and subset of probes with average count > 100 per replicate 
    X <- getHTTrDEG('DTXSID7020182',mncnt0=100)

  ... above and just subset of probes unadjusted p-value <= 0.05   
    X <- getHTTrDEG('DTXSID7020182',mncnt0=100,p0=0.05)

  ... above and just subset of probes unadjusted |l2fc0| >= 2   
    X <- getHTTrDEG('DTXSID7020182',mncnt0=100,p0=0.05,l2fc0=2)

  ... above and pivot on gene symbol and take the max of l2fc values    
    X <- getHTTrDEG('DTXSID7020182',mncnt0=100,p0=0.1,l2fc0=0.6,
                     pivot_on='gene', pivot_func='max')

}
