% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr_functions.R
\name{getHTTrPlateInfo}
\alias{getHTTrPlateInfo}
\title{Get the  HTTr plate information for a plate_id}
\usage{
getHTTrPlateInfo(plate_id)
}
\arguments{
\item{plate_id}{A valid plate_id (e.g. from getHTTrChemPlates)}
}
\value{
A dataframe with
}
\description{
Get the  HTTr plate information for a plate_id
}
\examples{
 getHTTrPlateInfo('TC00284691')      
}
