% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr_functions.R
\name{searchHTTrChem}
\alias{searchHTTrChem}
\title{Finds HTTr chemicals}
\usage{
searchHTTrChem(txt)
}
\arguments{
\item{txt}{A text/regex string}
}
\value{
a dataframe containing the chemical name, dsstox_sid and epa_sample_id
}
\description{
`searchHTTrChem` returns the HTTr chemicals than match the input string pattern.
}
\examples{
searchHTTrChem('phtha')                                  
}
